// src/types/partials/Coding.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TUri } from '../simpleTypes/Uri';

export type TCoding = {
  id?: string;
  extension?: TExtension[];
  system?: TUri;
  version?: string;
  code?: string;
  display?: string;
  userSelected?: boolean;
};